Move thumbnail helpers to thumbnail utility
